LPD-96725 Reuse the context data injector for nested fields
Build a single injector instead of rebuilding one per item on the shared builder, and move getAPIVersion back to the internal NestedFieldsSetterUtil so its reflective contract is not frozen as public API.
